pyproject.toml
[project]
name = "firedrake"
# <year>.<month>.<patch>
# TODO RELEASE
version = "2025.11.0.dev0"
description = "An automated system for the portable solution of partial differential equations using the finite element method"
readme = "README.rst"
license = "LGPL-3.0-or-later"
maintainers = [
  {name = "Pablo D. Brubeck"},
  {name = "Daiane I. Dolci"},
  {name = "David A. Ham", email = "david.ham@imperial.ac.uk"},
  {name = "Josh Hope-Collins"},
  {name = "Koki Sagiyama"},
  {name = "Connor J. Ward", email = "c.ward20@imperial.ac.uk"},
]
requires-python = ">=3.10"
dependencies = [
  "cachetools",
  "decorator<=4.4.2",
  "mpi4py>3; python_version >= '3.13'",
  "mpi4py; python_version < '3.13'",
  # TODO RELEASE
  "fenics-ufl @ git+https://github.com/FEniCS/ufl.git@main",
  # TODO RELEASE
  "firedrake-fiat @ git+https://github.com/firedrakeproject/fiat.git@main",
  "fuse-element @ git+https://github.com/indiamai/fuse.git",
  "h5py>3.12.1",
  "immutabledict",
  "libsupermesh",
  "loopy>2024.1",
  "numpy",
  "packaging",
  # TODO RELEASE
  # "petsc4py==3.24.0",
  # UNDO ME
  "petsctools @ git+https://github.com/firedrakeproject/petsctools.git@connorjward/cpetsc",
  "pkgconfig",
  "progress",
  # TODO RELEASE
  "pyadjoint-ad @ git+https://github.com/dolfin-adjoint/pyadjoint@master",
  "pycparser",
  "pytools[siphash]",
  "requests",
  "rtree>=1.2",
  "scipy",
  "sympy",
  "islpy>=2025.1.5; sys_platform == 'darwin'",
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: Unix",
]

[project.urls]
Homepage = "https://firedrakeproject.org"
Repository = "https://github.com/firedrakeproject/firedrake"
Issues = "https://github.com/firedrakeproject/firedrake/issues/new/choose"

[project.scripts]
firedrake-clean = "firedrake.scripts.firedrake_clean:main"
firedrake-preprocess-bibtex = "firedrake.scripts.firedrake_preprocess_bibtex:main"
pyop2-clean = "pyop2.compilation:clear_compiler_disk_cache"


[project.optional-dependencies]
check = [
  # TODO RELEASE
  "mpi-pytest @ git+https://github.com/firedrakeproject/mpi-pytest.git@main",
  "pytest",
]
docs = [
  "bibtexparser",
  "matplotlib",  # needed to resolve API
  "numpydoc",
  "pylit",
  "sphinx",
  "sphinx-autobuild",
  "sphinx-reredirects",
  "sphinxcontrib-bibtex",
  "sphinxcontrib-jquery",
  "sphinxcontrib-svg2pdfconverter",
  "sphinxcontrib-youtube",
  "sphinx-copybutton",
  "vtk",  # needed to resolve API
]
jax = [
  "jax",
]
# TODO RELEASE
# netgen = [
#   "ngsPETSc>=0.1.1",
# ]
# TODO RELEASE
# slepc = [
#   "slepc4py==3.24.0",
# ]
torch = [  # requires passing '--extra-index-url' to work
  "torch",
]
vtk = [
  "vtk",
]

# Dependencies needed to run the full test suite
ci = [
  "ipympl",  # needed for notebook testing
  "jax",
  "matplotlib",
  # TODO RELEASE
  "mpi-pytest @ git+https://github.com/firedrakeproject/mpi-pytest.git@main",
  "nbval",
  "networkx",
  # TODO RELEASE
  # "ngsPETSc>=0.1.1",
  "pdf2image",
  "pygraphviz",
  "pylit",
  "pytest",
  "pytest-order",
  "pytest-split",  # needed for firedrake-run-split-tests
  "pytest-timeout",
  "pytest-xdist",
  # TODO RELEASE
  # "slepc4py==3.24.0",
  "torch",  # requires passing '--extra-index-url' to work
  "vtk",
]
docker = [  # Used in firedrake-vanilla container
  "ipympl",  # needed for notebook testing
  "matplotlib",
  # TODO RELEASE
  "mpi-pytest @ git+https://github.com/firedrakeproject/mpi-pytest.git@main",
  "nbval",
  "networkx",
  "pdf2image",
  "pygraphviz",
  "pylit",
  "pytest",
  "pytest-order",
  "pytest-split",  # needed for firedrake-run-split-tests
  "pytest-timeout",
  "pytest-xdist",
  # TODO RELEASE
  # "slepc4py==3.24.0",
  "vtk",
]

[build-system]
requires = [
  "Cython>=3.0",
  "libsupermesh",
  "mpi4py>3; python_version >= '3.13'",
  "mpi4py; python_version < '3.13'",
  "numpy",
  # TODO RELEASE
  # "petsc4py==3.24.0",
  "petsctools",
  "pkgconfig",
  "pybind11",
  "rtree>=1.2",
  "setuptools>=77.0.3",
]
build-backend = "setuptools.build_meta"

# TODO: Convert firedrake-zenodo to a proper entrypoint script.
[tool.setuptools]
script-files = [
  "firedrake/scripts/firedrake-zenodo",
  "scripts/firedrake-check",
  "scripts/firedrake-run-split-tests",
]

[tool.setuptools.package-data]
# Unless specified these files will not be installed along with the
# rest of the package
firedrake = [
  "cython/*.pxi",
  "cython/*.pyx",
  "evaluate.h",
  "locate.c",
  "icons/*.png",
  "_check/**",
]
pyop3 = [
  "*.pyx",
  "ir/*.c",
]
